fixed send issues when not synced

+++ scenes/sendScene.js
@@ -5,10 +5,12 @@ const { getWalletById } = require("../utils/walletUtils");
 const step1 = async (ctx) => {
   const wallet = await getWalletById(ctx.session.loggedInWalletId);
   if (wallet.state.status !== "ready") {
-    try {
-      await ctx.deleteMessage();
-    } catch (error) {
-      console.log(error);
+    if (ctx.update.callback_query?.data === "refresh-send") {
+      try {
+        await ctx.deleteMessage();
+      } catch (error) {
+        console.log(error);
+      }
     }
     await ctx.reply(
       `Your wallet is syncing.
@@ -16,11 +18,11 @@ Please Wait for it to finish before trying to send...
         
 Progress: ${wallet.state.progress.quantity} ${wallet.state.progress.unit}`,
       Markup.inlineKeyboard([
-        [Markup.button.callback("Refresh", "send")],
+        [Markup.button.callback("Refresh", "refresh-send")],
         [mainMenuButton()],
       ])
     );
-    return;
+    return ctx.scene.leave();
   }
   ctx.reply(
     "Please choose an option from the choices below",
